server/pyproject.toml
[project]
name = "transit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.5",
    "httpx>=0.28.1",
    "python-decouple>=3.8",
]
